requirements.txt
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
bleach==6.0.0
botocore==1.29.76
build==0.10.0
cached-property==1.5.2
cachetools==5.3.1
Cerberus==1.3.4
certifi==2023.5.7
chardet==5.0.0
charset-normalizer==3.1.0
colorama==0.4.6
dataclasses-json==0.5.8
decorator==5.1.1
distlib==0.3.6
docutils==0.20.1
executing==1.2.0
frozenlist==1.3.3
fsspec==2023.6.0
google-api-core==2.11.0
google-api-python-client==2.89.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
ipython==8.14.0
jaraco.classes==3.2.3
jedi==0.18.2
jmespath==1.0.1
keyring==23.13.1
langchain==0.0.198
langchainplus-sdk==0.0.9
llama-index==0.6.25.post1
markdown-it-py==3.0.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==9.1.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
orderedmultidict==1.0.1
packaging==23.1
pandas==2.0.2
parso==0.8.3
pep517==0.13.0
pexpect==4.8.0
pickleshare==0.7.5
pip-shims==0.7.3
pipenv-setup==3.2.0
pipfile==0.0.2
pkginfo==1.9.6
platformdirs==3.5.3
plette==0.4.4
progress==1.6
prompt-toolkit==3.0.38
protobuf==4.23.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
Pygments==2.15.1
pyparsing==3.0.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
readme-renderer==37.3
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
requirementslib==1.6.9
rfc3986==2.0.0
rich==13.4.2
rsa==4.9
s3fs==2023.6.0
six==1.16.0
SQLAlchemy==2.0.16
stack-data==0.6.2
tenacity==8.2.2
tiktoken==0.4.0
toml==0.10.2
tomlkit==0.11.8
tqdm==4.65.0
traitlets==5.9.0
twine==4.0.2
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vistir==0.8.0
wcwidth==0.2.6
webencodings==0.5.1
wrapt==1.15.0
yarl==1.9.2
youtube-transcript-api==0.6.0
zipp==3.15.0
